refactor(react-ui): consolidate MarkdownMedia into MediaPlayer
Fold the iframe + image fallback branches from `MarkdownMedia` into
`MediaPlayer` so a single component dispatches between native
`<video>`/`<audio>`, oEmbed-style `<iframe>` (Cloudflare Stream, etc.),
and `<img>` based on the URL. Drop `react-ui-markdown/MarkdownView/Media.tsx`
and update `MarkdownView` and `plugin-support` consumers to use
`MediaPlayer` from `@dxos/react-ui` directly.
