+++ packages/ui/react-ui-markdown/src/MarkdownView/MarkdownView.tsx
@@ -6,11 +6,9 @@ import React, { type PropsWithChildren } from 'react';
 import ReactMarkdown, { type Options as ReactMarkdownOptions } from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 
-import { type ThemedClassName } from '@dxos/react-ui';
+import { MediaPlayer, type ThemedClassName } from '@dxos/react-ui';
 import { SyntaxHighlighter } from '@dxos/react-ui-syntax-highlighter';
 import { mx } from '@dxos/ui-theme';
-
-import { MarkdownMedia } from './Media';
 
 export type MarkdownViewProps = ThemedClassName<
   PropsWithChildren<{
@@ -73,13 +71,13 @@ const defaultComponents: ReactMarkdownOptions['components'] = {
   // leaving the browser's broken-image placeholder.
   //
   // Media URLs (mp4/mp3/etc. or legacy `iframe`-style embeds) are swapped to a
-  // native `<video>` / `<audio>` `MediaPlayer` by {@link MarkdownMedia} so
-  // playable media in the source document renders inline.
+  // native `<video>` / `<audio>` element by {@link MediaPlayer} so playable
+  // media in the source document renders inline.
   img: ({ src, alt }) => {
     if (!src) {
       return null;
     }
-    return <MarkdownMedia src={src} alt={alt} mediaClassNames='aspect-video w-full' />;
+    return <MediaPlayer src={src} alt={alt} mediaClassNames='aspect-video w-full' />;
   },
   ol: ({ children, ...props }) => (
     <ol className='pt-1 pb-1 ps-6 leading-tight list-decimal' {...props}>
