+++ packages/ui/react-ui-markdown/src/MarkdownView/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './MarkdownView';
-export * from './Media';
